refactor(e2e): rename and restructure E2E test classes for products and members
- Moved `ProductControllerTest` and `MemberControllerTest` to the `ecommerce.etoe` package.
- Renamed classes to `ProductE2ETest` and `MemberE2ETest`, respectively, for better alignment with their purpose.

refactor(member): rename DTO mapper methods for clarity

- Renamed `toDto` to `toResponseDto` across `MemberMapper` and dependent service implementations.
- Improved method naming to convey the intent of converting to a response format.

refactor(product): use snake_case for database column names

- Updated `ProductRepositoryImpl` to use `image_url` instead of `imageUrl` in queries and mappings.

feat(controller): add role-based access control annotations

- Introduced `@CheckAdminOnly` and `@IgnoreCheckLogin` annotations in `ProductController` for method-level access control.
