+++ src/main/kotlin/ecommerce/controller/ProductController.kt
@@ -1,5 +1,7 @@
 package ecommerce.controller
 
+import ecommerce.annotation.CheckAdminOnly
+import ecommerce.annotation.IgnoreCheckLogin
 import ecommerce.model.ProductDTO
 import ecommerce.model.ProductPatchDTO
 import ecommerce.services.ProductService
@@ -17,14 +19,17 @@ import java.net.URI
 
 @RestController
 class ProductController(private val productService: ProductService) {
+    @IgnoreCheckLogin
     @GetMapping(PRODUCT_PATH)
     fun getProducts(): List<ProductDTO> = productService.findAll()
 
+    @IgnoreCheckLogin
     @GetMapping(PRODUCT_PATH_ID)
     fun getProductById(
         @PathVariable id: Long,
     ): ResponseEntity<ProductDTO> = ResponseEntity.ok(productService.findById(id))
 
+    @CheckAdminOnly
     @PostMapping(PRODUCT_PATH)
     fun createProduct(
         @Valid @RequestBody productDTO: ProductDTO,
@@ -33,18 +38,21 @@ class ProductController(private val productService: ProductService) {
         return ResponseEntity.created(URI.create("$PRODUCT_PATH/${saved.id}")).body(saved)
     }
 
+    @CheckAdminOnly
     @PutMapping(PRODUCT_PATH_ID)
     fun updateProductById(
         @Valid @RequestBody productDTO: ProductDTO,
         @PathVariable id: Long,
     ): ResponseEntity<ProductDTO> = ResponseEntity.ok(productService.updateById(id, productDTO))
 
+    @CheckAdminOnly
     @PatchMapping(PRODUCT_PATH_ID)
     fun patchProductById(
         @Valid @RequestBody productPatchDTO: ProductPatchDTO,
         @PathVariable id: Long,
     ): ResponseEntity<ProductDTO> = ResponseEntity.ok(productService.patchById(id, productPatchDTO))
 
+    @CheckAdminOnly
     @DeleteMapping(PRODUCT_PATH_ID)
     fun deleteProductById(
         @PathVariable id: Long,
@@ -53,6 +61,7 @@ class ProductController(private val productService: ProductService) {
         return ResponseEntity.noContent().build()
     }
 
+    @CheckAdminOnly
     @DeleteMapping(PRODUCT_PATH)
     fun deleteAllProducts(): ResponseEntity<String> {
         productService.deleteAll()
