+++ src/main/kotlin/ecommerce/mappers/MemberMapper.kt
@@ -3,6 +3,6 @@ package ecommerce.mappers
 import ecommerce.entities.Member
 import ecommerce.model.MemberDTO
 
-fun Member.toDto() = MemberDTO(id, email, password, role)
+fun Member.toResponseDto() = MemberDTO(id, email, password, role)
 
 fun MemberDTO.toEntity() = Member(id, email, password, role)
