+++ src/test/kotlin/ecommerce/etoe/MemberE2ETest.kt
@@ -1,4 +1,4 @@
-package ecommerce.controller
+package ecommerce.etoe
 
 import ecommerce.model.MemberDTO
 import ecommerce.model.TokenRequestDTO
@@ -17,7 +17,7 @@ import org.springframework.test.annotation.DirtiesContext
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class MemberControllerTest {
+class MemberE2ETest {
     lateinit var token: String
 
     @BeforeEach
